requirements.txt
Django==1.8.9
django-registration==2.0.4
Pillow==3.1.1